+++ packages/kobber-audio-recorder/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gyldendal/kobber-audio-recorder
 
+## 0.3.48
+
+### Patch Changes
+
+- Add Carousel exports
+
 ## 0.3.47
 
 ### Patch Changes