+++ packages/kobber-icons/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gyldendal/kobber-components
 
+## 0.3.48
+
+### Patch Changes
+
+- Add Carousel exports
+- Updated dependencies
+  - @gyldendal/kobber-base@0.3.48
+
 ## 0.3.47
 
 ### Patch Changes