+++ packages/kobber-prettier/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.0.1
 
+## 0.3.48
+
+### Patch Changes
+
+- Add Carousel exports
+
 ## 0.3.47
 
 ### Patch Changes